public url available, minor refactoring + socketio specific channel for boards

+++ backend/app.py
@@ -1,21 +1,22 @@
-from dotenv import load_dotenv
-load_dotenv()
-
-import uvicorn, socketio
-import os, asyncio
+import uvicorn
+import socketio
+import os
+import asyncio
 from fastapi import FastAPI, HTTPException, Depends, APIRouter
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from fastapi.responses import FileResponse
 from contextlib import asynccontextmanager
 from jose import jwt
 from bson import ObjectId, Binary
-from models import *
-import uuid, time
+from models import Role, Category, Product, Member, UserDTO, AddUser
+from models import BoardDTO, AddBoardForm, AddCategory, AddMember, AddProduct, IdResponse
+import uuid
+import time
 from fastapi.middleware.cors import CORSMiddleware
-
-from utils import crypto
-from env import DEBUG, CORS_ALLOW, JWT_ALGORITHM, APP_SECRET, JWT_EXPIRE_H
-from db import Role, init_db, shutdown_db, User, first_run, Board
-from fastapi.responses import FileResponse
+from env import DEBUG, CORS_ALLOW, JWT_ALGORITHM, JWT_EXPIRE_H
+from utils import APP_SECRET
+from db import init_db, shutdown_db, User, first_run, Board
+from db import crypto
 
 redis_mgr = socketio.AsyncRedisManager(
     url="redis://localhost:6379/0" if DEBUG else "redis://redis:6379/0",
@@ -26,7 +27,7 @@ sio_app = socketio.ASGIApp(sio_server, socketio_path="")
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     await init_db()
-    await front_refresh()
+    await sio_server.emit("update", "init")
     yield
     await shutdown_db()
 
@@ -38,9 +39,6 @@ async def sio_connect(sid, environ): pass
 
 @sio_server.on("disconnect")
 async def sio_disconnect(sid): pass
-
-async def front_refresh(additional:list[str]|None=None):
-    await sio_server.emit("update",[] if additional is None else additional)
 
 async def create_access_token(data: dict):
     to_encode = data.copy()
@@ -66,7 +64,7 @@ def has_role(target:Role|None = None):
         if target is None or auth == Role.admin:
             return True
         if target == Role.guest:
-            if not auth is None:
+            if auth is not None:
                 return True
         if target == Role.editor:
             if auth in [Role.editor, Role.admin]:
@@ -111,33 +109,41 @@ async def get_boards():
 async def new_board(form: AddBoardForm):
     """ Create a new board """
     board:Board = await Board(**form.model_dump(), categories=[], members=[], products=[]).save()
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return { "id": board.id.binary.hex() }
 
 @editor_api.delete("/boards/{id}", response_model=IdResponse, tags=["board"])
 async def remove_board(id: str):
     """ Create a new board """
     await Board.find_one(Board.id == ObjectId(id)).delete_one()
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return { "id": id }
 
-@guest_api.get("/boards/{id}", response_model=BoardDTO, tags=["board"])
+@api.get("/boards/{id}", response_model=BoardDTO, tags=["board"])
 async def get_board(id: str):
     """ Get board """
-    return await Board.find_one(Board.id == ObjectId(id))
+    board = await Board.find_one(Board.id == ObjectId(id))
+    if board is None:
+        raise HTTPException(404, "Board not found")
+    return board
 
 @editor_api.post("/boards/{id}", response_model=IdResponse, tags=["board"])
 async def edit_board(id: str, form: AddBoardForm):
     """ Edit board """
     board = await Board.find_one(Board.id == ObjectId(id))
+    if board is None:
+        raise HTTPException(404, "Board not found")
     await board.set(form)
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return { "id": id }
 
-@guest_api.get("/boards/{id}/categories", response_model=list[Category], tags=["category"])
+@api.get("/boards/{id}/categories", response_model=list[Category], tags=["category"])
 async def get_board_categories(id: str):
     """ Get board category list """
-    return (await Board.find_one(Board.id == ObjectId(id))).categories
+    board = await Board.find_one(Board.id == ObjectId(id))
+    if board is None:
+        raise HTTPException(404, "Board not found")
+    return board.categories
 
 
 @editor_api.put("/boards/{id}/categories", response_model=IdResponse, tags=["category"])
@@ -147,7 +153,7 @@ async def new_board_categories(id: str, form: AddCategory):
     await Board.find_one(Board.id == ObjectId(id)).update_one(
         {"$push": { "categories": Category(**form.model_dump(), id=new_id).model_dump() }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(new_id)}
 
 @editor_api.post("/boards/{id}/categories/{category_id}", response_model=IdResponse, tags=["category"])
@@ -158,7 +164,7 @@ async def edit_board_categories(id: str, category_id: str, form: AddCategory):
         {"$set": mongo_dict_update("categories.$[cat]", form.model_dump())},
         array_filters=[ {"cat.id": Binary.from_uuid(category_id)}]
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(category_id)}
 
 @editor_api.delete("/boards/{id}/categories/{category_id}", response_model=IdResponse, tags=["category"])
@@ -172,13 +178,16 @@ async def delete_board_categories(id: str, category_id: str):
             "members.$[].categories": Binary.from_uuid(category_id),
         }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(category_id)}
 
-@guest_api.get("/board/{id}/members", response_model=list[Member], tags=["member"])
+@api.get("/boards/{id}/members", response_model=list[Member], tags=["member"])
 async def get_board_members(id: str):
     """ Get board member list """
-    return (await Board.find_one(Board.id == ObjectId(id))).members
+    board = await Board.find_one(Board.id == ObjectId(id))
+    if board is None:
+        raise HTTPException(404, "Board not found")
+    return board.members
 
 @editor_api.put("/boards/{id}/members", response_model=IdResponse, tags=["member"])
 async def new_board_members(id: str, form: AddMember):
@@ -187,7 +196,7 @@ async def new_board_members(id: str, form: AddMember):
     await Board.find_one(Board.id == ObjectId(id)).update_one(
         {"$push": { "members": Member(**form.model_dump(), id=new_id).model_dump() }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(new_id)}
 
 @editor_api.post("/boards/{id}/members/{member_id}", response_model=IdResponse, tags=["member"])
@@ -198,7 +207,7 @@ async def edit_board_members(id: str, member_id: str, form: AddMember):
         {"$set": mongo_dict_update("members.$[memb]", form.model_dump()) },
         array_filters=[ {"memb.id": Binary.from_uuid(member_id)}]
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(member_id)}
 
 @editor_api.delete("/boards/{id}/members/{member_id}", response_model=IdResponse, tags=["member"])
@@ -208,13 +217,16 @@ async def delete_board_members(id: str, member_id: str):
     await Board.find_one(Board.id == ObjectId(id)).update_one(
         {"$pull": { "members": { "id": Binary.from_uuid(member_id)} }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(member_id)}
 
-@guest_api.get("/boards/{id}/products", response_model=list[Product], tags=["product"])
+@api.get("/boards/{id}/products", response_model=list[Product], tags=["product"])
 async def get_board_products(id: str):
     """ Get board product list """
-    return (await Board.find_one(Board.id == ObjectId(id))).products
+    board = await Board.find_one(Board.id == ObjectId(id))
+    if board is None:
+        raise HTTPException(404, "Board not found")
+    return board.products
 
 @editor_api.put("/boards/{id}/products", response_model=IdResponse, tags=["product"])
 async def new_board_products(id: str, form: AddProduct):
@@ -223,7 +235,7 @@ async def new_board_products(id: str, form: AddProduct):
     await Board.find_one(Board.id == ObjectId(id)).update_one(
         {"$push": { "products": Product(**form.model_dump(), id=new_id).model_dump() }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(new_id)}
 
 @editor_api.post("/boards/{id}/products/{product_id}", response_model=IdResponse, tags=["product"])
@@ -234,7 +246,7 @@ async def edit_board_products(id: str, product_id: str, form: AddProduct):
         {"$set": mongo_dict_update("products.$[prod]", form.model_dump())},
         array_filters=[ {"prod.id": Binary.from_uuid(product_id)}]
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(product_id)}
 
 @editor_api.delete("/boards/{id}/products/{product_id}", response_model=IdResponse, tags=["product"])
@@ -244,7 +256,7 @@ async def delete_board_products(id: str, product_id: str):
     await Board.find_one(Board.id == ObjectId(id)).update_one(
         {"$pull": { "products": { "id": Binary.from_uuid(product_id)} }}
     )
-    await front_refresh()
+    await sio_server.emit(f"update:{id.lower()}", "update")
     return {"id":str(product_id)}
 
 @admin_api.get("/users", response_model=list[UserDTO], tags=["user"])
@@ -255,7 +267,10 @@ async def get_users():
 @admin_api.get("/users/{id}", response_model=UserDTO, tags=["user"])
 async def get_user(id: str):
     """ Get user """
-    return await User.find_one(User.id == ObjectId(id))
+    user = await User.find_one(User.id == ObjectId(id))
+    if user is None:
+        raise HTTPException(404, "User not found")
+    return user
 
 @admin_api.put("/users", response_model=IdResponse, tags=["user"])
 async def new_user(form: AddUser):
@@ -273,7 +288,7 @@ async def new_user(form: AddUser):
         )
     form.password=crypto.hash(form.password)
     user:User = await User(**form.model_dump()).save()
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return {"id": user.id.binary.hex()}
 
 @admin_api.post("/users/{id}", response_model=IdResponse, tags=["user"])
@@ -288,21 +303,25 @@ async def edit_user(id: str, form: AddUser):
     if form.password:
         form.password = crypto.hash(form.password)
     user = await User.find_one(User.id == ObjectId(id))
+    if user is None:
+        raise HTTPException(404, "User not found")
     await user.set(form)
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return {"id":id}
 
 @admin_api.delete("/users/{id}", response_model=IdResponse, tags=["user"])
 async def delete_users(id: str):
     """ Delete a user """
     user = await User.find_one(User.id == ObjectId(id))
+    if user is None:
+        raise HTTPException(404, "User not found")
     if user.username == "admin":
         raise HTTPException(
             status_code=400,
             detail="'admin' is reserved"
         )
     await User.find_one(User.id == ObjectId(id)).delete_one()
-    await front_refresh()
+    await sio_server.emit("update", "update")
     return {"id":id}
 
 app.include_router(api)
